# xs2a/xs2a-api/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Reference Implementation of XS2A REST Services",
    "version" : "0.9.2-SNAPSHOT",
    "title" : "XS2A Services",
    "contact" : {
      "name" : "Andreas Reetz",
      "email" : "andreas.reetz@sdv-it.de"
    }
  },
  "host" : "api.sdv-it.de",
  "basePath" : "/tpp",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/{version}/consents/{consent-id}" : {
      "get" : {
        "tags" : [ "Consents" ],
        "summary" : "Get Consent Request (page 56) - returns the content of of an account information consent object",
        "description" : "",
        "operationId" : "consentRequest",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/ConsentRequestResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      },
      "put" : {
        "tags" : [ "Consents" ],
        "summary" : "Update PSU Data (page 70) - updates the account information consent data by PSU data",
        "description" : "",
        "operationId" : "updatePSUData",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only in corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UpdatePSUDataRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/UpdatePSUDataResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      },
      "delete" : {
        "tags" : [ "Consents" ],
        "summary" : "Delete an Account Information Consent Object (page 58) - deletes the consent object",
        "description" : "",
        "operationId" : "deleteConsent",
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "OK, request received"
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/consents/{consent-id}/status" : {
      "get" : {
        "tags" : [ "Consents" ],
        "summary" : "Get Status Request (page 55) - reads the transaction status of the addressed consent resource",
        "description" : "",
        "operationId" : "statusRequest",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/StatusResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/payments/{payment-product}" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Payment Initiation Request (page 25) - creates a payment initiation request at the ASPSP",
        "description" : "",
        "operationId" : "paymentInitiationRequest",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-product",
          "in" : "path",
          "description" : "sepa-credit-transfers / instant-sepa-credit-transfers / target-2-payments / cross-border-credit-transfers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Consent-ID",
          "in" : "header",
          "description" : "May be contained, if the payment initiation transaction is part of a combined AIS/PIS service",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Agent",
          "in" : "header",
          "description" : "The forwarded Agent header field of the http request between PSU and TPP",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-IP-Address",
          "in" : "header",
          "description" : "The forwarded IP Address header field consists of the corresponding http request IP Address between PSU and TPP",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-Geo-Location",
          "in" : "header",
          "description" : "The forwarded Geo Location header field of the corresponding http request between PSU and TPP if available",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PaymentInitiationRequest"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/PaymentInitiationResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
...
    "/{version}/payments/{payment-product}/{payment-id}/status" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Get Status Request (page 33) - check the status of the payment initiation",
        "description" : "",
        "operationId" : "statusRequest",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-product",
          "in" : "path",
          "description" : "sepa-credit-transfers / instant-sepa-credit-transfers / target-2-payments / cross-border-credit-transfers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-id",
          "in" : "path",
          "description" : "Payment ID as received from the Payment Initiation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/StatusResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ConsentRequestResponse" : {
      "type" : "object",
      "required" : [ "access_accounts", "consent_status", "frequency_per_day", "recurring_indicator", "transaction_status", "valid_until" ],
      "properties" : {
        "access_accounts" : {
          "type" : "array",
          "description" : "List of accounts to be accessed",
          "items" : {
            "$ref" : "#/definitions/ConsentAccessAccount"
          }
        },
        "recurring_indicator" : {
          "type" : "boolean",
          "example" : true,
          "description" : "\"true\" if the consent is for requring access"
        },
        "valid_until" : {
          "type" : "string",
          "example" : "2017-10-30",
          "description" : "Valid until date for the requested consent"
        },
        "frequency_per_day" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 4,
          "description" : "Maximum frequency for an access per day"
        },
        "transaction_status" : {
          "type" : "string",
          "example" : "AcceptedTechnicalValidation",
          "description" : "Transaction status"
        },
        "consent_status" : {
          "type" : "string",
          "example" : "valid",
          "description" : "Only permitted: empty|valid|blocked|expired|deleted"
        }
      }
    },
...
    "PSUData" : {
      "type" : "object",
      "properties" : {
        "password" : {
          "type" : "string",
          "description" : "PSU password for Update PSU Data (Authentication)"
        }
      }
    },
    "PaymentInitiationRequest" : {
      "type" : "object",
      "required" : [ "creditor", "creditor_account", "debtor_account", "instructed_amount" ],
      "properties" : {
        "instructed_amount" : {
          "description" : "currency and amount",
          "$ref" : "#/definitions/AccountAmount"
        },
        "debtor_account" : {
          "description" : "Debtor account",
          "$ref" : "#/definitions/AccountAccount"
        },
        "creditor" : {
          "example" : "Merchant123",
          "description" : "Creditor name",
          "$ref" : "#/definitions/AccountName"
        },
        "creditor_account" : {
          "description" : "Creditor account",
          "$ref" : "#/definitions/AccountAccount"
        },
        "remittance_information_structured" : {
          "type" : "string",
          "example" : "false",
          "description" : "\"true\": a payment initiation service will be address in the same \"session\""
        }
      }
    },
...
    "StatusResponse" : {
      "type" : "object",
      "required" : [ "transaction_status" ],
      "properties" : {
        "transaction_status" : {
          "type" : "string",
          "description" : "Status"
        }
      }
    },
...
    "UpdatePSUDataRequest" : {
      "type" : "object",
      "properties" : {
        "authentication_method_id" : {
          "type" : "string",
          "description" : "authentication method id for Update PSU Data (Authentication Method)"
        },
        "psu_data" : {
          "description" : "PSU Data for Update PSU Data (Authentication)",
          "$ref" : "#/definitions/PSUData"
        },
        "sca_authentication_data" : {
          "type" : "string",
          "description" : "sca_authentication_data - for Authorise Request"
        }
      },
      "description" : "Update PSU Data Request"
    },
    "UpdatePSUDataResponse" : {
      "type" : "object",
      "required" : [ "_links", "transaction_status" ],
      "properties" : {
        "sca_methods" : {
          "type" : "array",
          "description" : "name or names of autheticated methods",
          "items" : {
            "$ref" : "#/definitions/AuthenticationMethod"
          }
        },
        "chosen_sca_method" : {
          "description" : "name of chosen method",
          "$ref" : "#/definitions/AuthenticationMethod"
        },
        "sca_chalenge_data" : {
          "description" : "Extended information to chosen_sca_method",
          "$ref" : "#/definitions/AuthenticationChallengeData"
        },
        "transaction_status" : {
          "type" : "string",
          "description" : "Status"
        },
        "psu_message" : {
          "type" : "string",
          "description" : "Text to be displayed to the PSU"
        },
        "_links" : {
          "description" : "Links",
          "$ref" : "#/definitions/Links"
        }
      }
    }
  }
}